Refactor db code
