+++ core/database/eventapplicant.go
@@ -1,8 +1,6 @@
 package database
 
 import (
-	"strings"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
@@ -16,134 +14,6 @@ const (
 	applicantPrefix = "Applicant-"
 	tagPrefix       = "Tag-"
 )
-
-type eventItem struct {
-	PeidPK      string          `json:"pk"`
-	PeidSK      string          `json:"sk"`
-	Name        string          `json:"name"`
-	Description string          `json:"description"`
-	Fields      []*models.Field `json:"fields"`
-}
-
-type applicantItem struct {
-	Period string   `json:"pk"`
-	Email  string   `json:"sk"`
-	Name   string   `json:"name"`
-	Tags   []string `json:"tags"`
-}
-
-type applicationItem struct {
-	Peid    string                        `json:"pk"`
-	Email   string                        `json:"sk"`
-	Name    string                        `json:"name"`
-	Entires map[string]*models.FieldEntry `json:"entries"`
-}
-
-type tagItem struct {
-	Period  string `json:"pk"`
-	TagName string `json:"sk"`
-}
-
-func newDBEvent(event *models.Event) *eventItem {
-	peid := prefixPeriodEventID(event.Period, event.EventID)
-	return &eventItem{
-		PeidPK:      peid,
-		PeidSK:      peid,
-		Name:        event.Name,
-		Description: event.Description,
-		Fields:      event.Fields,
-	}
-}
-
-func getEvent(item *eventItem) *models.Event {
-	p, e := getPeriodAndEventID(item.PeidPK)
-	return &models.Event{
-		Period:      p,
-		EventID:     e,
-		Name:        item.Name,
-		Description: item.Description,
-		Fields:      item.Fields,
-	}
-}
-
-func newDBApplicant(applicant *models.Applicant) *applicantItem {
-	p := prefixPeriodID(applicant.Period)
-	return &applicantItem{
-		Period: p,
-		Email:  prefixApplicantID(applicant.Email),
-		Name:   applicant.Name,
-		Tags:   applicant.Tags,
-	}
-}
-
-func getApplicant(item *applicantItem) *models.Applicant {
-	return &models.Applicant{
-		Period: removePrefix(item.Period),
-		Email:  removePrefix(item.Email),
-		Name:   item.Name,
-		Tags:   item.Tags,
-	}
-}
-
-func newDBTag(tag *models.Tag) *tagItem {
-	return &tagItem{
-		Period:  prefixPeriodID(tag.Period),
-		TagName: prefixTag(tag.TagName),
-	}
-}
-
-func getTag(item *tagItem) *models.Tag {
-	return &models.Tag{
-		Period:  removePrefix(item.Period),
-		TagName: removePrefix(item.TagName),
-	}
-}
-
-func newDBApplication(application *models.Application) *applicationItem {
-	peid := prefixPeriodEventID(application.Period, application.EventID)
-	return &applicationItem{
-		Peid:    peid,
-		Email:   prefixApplicantID(application.Email),
-		Name:    application.Name,
-		Entires: application.Entires,
-	}
-}
-
-func getApplication(item *applicationItem) *models.Application {
-	p, e := getPeriodAndEventID(item.Peid)
-	return &models.Application{
-		Period:  p,
-		EventID: e,
-		Email:   removePrefix(item.Email),
-		Name:    item.Name,
-	}
-}
-
-func getClubTable(clubID string) *string {
-	return aws.String(clubTablePrefix + clubID)
-}
-
-func prefixPeriodEventID(period string, eventID string) string {
-	return peidPrefix + period + "-" + eventID
-}
-
-func prefixPeriodID(period string) string {
-	return periodPrefix + period
-}
-
-func prefixApplicantID(email string) string {
-	return applicantPrefix + email
-}
-
-func prefixTag(tag string) string {
-	return tagPrefix + tag
-}
-
-func getPeriodAndEventID(peid string) (string, string) {
-	str := removePrefix(peid)
-	s := strings.SplitN(str, "-", 2)
-	return s[0], s[1]
-}
 
 // AddNewEvent creates a new event in the club table
 func (db *Database) AddNewEvent(clubID string, event *models.Event) error {
